+++ src/build.rs
@@ -1,15 +1,19 @@
 use crate::{
     action::Entity,
     components::{
-        account::{new_account_details_component, new_account_list_component},
         entity_types::new_entity_types_list,
         fps::FpsCounter,
-        group::ComponentGroup,
-        layout::RootLayout,
+        group::{
+            group::ComponentGroup,
+            layout::RootLayout,
+            split::{Axis, SplitComponent},
+            switch::SwitchComponent,
+        },
+        list_and_details::{
+            account::{new_account_details_component, new_account_list_component},
+            utxo::{new_utxo_details_component, new_utxo_list_component},
+        },
         message::Message,
-        split::{Axis, SplitComponent},
-        switch::SwitchComponent,
-        utxo::{new_utxo_details_component, new_utxo_list_component},
     },
     focus::{FocusManager, FocusableComponent},
     shared::{Shared, shared},
