+++ src/components.rs
@@ -1,3 +1,4 @@
+use crate::{action::Action, config::Config, tui::Event};
 use color_eyre::Result;
 use crossterm::event::{KeyEvent, MouseEvent};
 use ratatui::{
@@ -6,21 +7,14 @@ use ratatui::{
 };
 use tokio::sync::mpsc::UnboundedSender;
 
-use crate::{action::Action, config::Config, tui::Event};
-
-pub mod account;
 pub mod details;
 pub mod empty;
 pub mod entity_types;
 pub mod fps;
 pub mod group;
 pub mod home;
-pub mod layout;
+pub mod list_and_details;
 pub mod message;
-pub mod scroll;
-pub mod split;
-pub mod switch;
-pub mod utxo;
 
 /// `Component` is a trait that represents a visual and interactive element of the user interface.
 ///
