+++ src/components/list_and_details.rs
@@ -1,0 +1,2 @@
+pub mod account;
+pub mod utxo;
